"Changing the callback invocation"

+++ SW/LCD/lcd1602A.h
@@ -105,7 +105,7 @@ void LCD_puts ( char *a );
 // Prints a string on the LCD that is allocated on the program memory
 void LCD_putrs ( const char *a );
 
-void LCD_putNum(uint8_t number);
+void LCD_putNum(uint32_t number);
 
 // [Internal use only] Write a byte to the LCD
 void LCD_Write ( unsigned char c );
